scripts/requirements.txt
pandas>=2.2.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
openpyxl>=3.1.0
python-dotenv>=1.0.0

